Cleanup
Redo the to_human_readable function.
This is probably a cleaner and simpler looking function now, I'm happy with this.

+++ src/utils.hpp
@@ -1,5 +1,4 @@
 #pragma once
 #include <string>
-#include <cstdint>
 
-std::string to_human_readable(const uint64_t&);
+std::string to_human_readable(const uint&);
